cdk/requirements.txt
aws-cdk-lib==2.195.0
nodejs